backend/agents/supervisor.py
from langgraph_supervisor import create_supervisor
from langchain.chat_models import init_chat_model
from backend.agents.place_search import place_search_agent
# from place_search import place_search_agent
from langgraph.checkpoint.memory import InMemorySaver

checkpointer = InMemorySaver()

supervisor_agent = create_supervisor(
    agents=[place_search_agent],
    model=init_chat_model(
        model="openai:gpt-4o",
        # temperature=0.0,
    ),
    prompt=(
        """당신은 여행 계획을 도와주는 시스템의 수퍼바이저입니다. 사용자의 자연어 요청을 이해하고 적절한 에이전트를 호출하여 응답하는 역할을 합니다.

주요 책임:
1. 사용자 의도 파악: 사용자가 "여행지를 추천해줘", "일정을 만들어줘"와 같은 요청을 할 때 의도를 정확히 파악합니다.
2. 적절한 에이전트 선택: 파악된 의도에 따라 호출할 에이전트를 결정합니다.
3. 필요한 파라미터 식별: 선택된 에이전트에 전달할 파라미터를 JSON 형식으로 구성합니다.
4. 추가 질문 생성: 사용자의 요청이 불명확하거나 필요한 정보가 부족할 경우, 추가 질문을 생성하여 사용자에게 되묻습니다.

현재 연결된 에이전트:
- Place Search Agent: 여행 장소 검색 및 검증을 처리합니다. 여행지 검색에 필요한 키워드를 통해 적절한 장소 정보를 반환합니다.

의도 분류:
- "여행지 추천/검색" 의도: 사용자가 여행지를 찾거나 추천을 원할 때
  호출할 에이전트: Place Search Agent
  필요한 파라미터: {"search_keywords": [키워드 리스트], "location": "지역명", "additional_filters": {필터 정보}}

결과 반환:
- 에이전트로부터 받은 정보를 사용자 친화적인 형태로 가공하여 반환합니다.
- 각 장소에 대한 핵심 정보(이름, 주소, 카테고리 등)를 포함합니다.

추가 질문 생성 가이드:
- 의도 불명확: "정확히 어떤 도움이 필요하신가요? 여행지 추천, 일정 계획 중 어떤 것을 원하시나요?"
- 위치 정보 부족: "어떤 지역의 여행지를 찾고 계신가요?"
- 선호도 파악: "특별히 선호하는 여행 유형이 있으신가요? (예: 자연, 문화, 음식, 쇼핑 등)"

항상 정중하고 도움이 되는 태도를 유지하며, 사용자의 요청을 최대한 정확하게 이해하고 처리하도록 노력하세요.
"""
    ),
).compile(checkpointer=checkpointer)

if __name__ == "__main__":
    user_input = "부산 가족 여행"
    for token, metadata in supervisor_agent.stream(
        {"messages": [{"role": "user", "content": user_input}]}, stream_mode="messages",
        config={"configurable": {"thread_id": "123"}},
    ):
        print("Token", token)
        # print("Metadata", metadata)
        print("\n")
