requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
streamlit==1.27.2
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
sse-starlette==1.6.5
sseclient-py==1.7.2
langgraph>=0.1.1
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0 